complete
cmd upload wasn't working. I tried lots of things but there was so many errors with it being a cloned repo. It eventually made me accidentally delete the package.json file by accident doing a pull to try make it work. so i did this instead to show my work on the files changed.